BOTS/discord-bot/requirements.txt
discord.py-self==1.9.2
aiohttp>=3.8.0
groq>=0.4.0
questionary>=2.0.0
rich>=13.0.0
nltk>=3.8.0
python-dotenv>=1.0.0
aiofiles>=23.0.0